some modifications of authentification

+++ MainWindow.h
@@ -51,5 +51,5 @@ class MainWindow : public QMainWindow, public Ui::MainWindow
 
         // This is the socket that will let us communitate with the server.
         QTcpSocket *socket;
-        int blocksize;
+        quint16 blocksize;
 };
